+++ classes/Address.class.php
@@ -82,20 +82,28 @@ class Address
      * `$data` may be an array or a json_encoded string.
      *
      * @param   string|array    $data   Address data
-     * @return  object          $this
      */
     public function __construct($data=array())
     {
-        global $_SHOP_CONF;
+        global $_SHOP_CONF, $_USER;
 
         if (!is_array($data)) {
             // Allow for a JSON string to be provided.
             $data = json_decode($data, true);
         }
-        if (!is_array($data)) {
-            $data = array();
+        if (is_array($data)) {
+            $this->setVars($data);
         }
-        return $this->setUid(SHOP_getVar($data, 'uid', 'integer'))
+        if ($this->uid < 1) {
+            // in case an empty object is being created, set the user ID
+            $this->setUid($_USER['uid']);
+        }
+    }
+
+
+    public function setVars($data)
+    {
+        $this->setUid(SHOP_getVar($data, 'uid', 'integer'))
             ->setID(SHOP_getVar($data, 'addr_id', 'integer'))
             ->setBilltoDefault(SHOP_getVar($data, 'billto_def', 'integer'))
             ->setShiptoDefault(SHOP_getVar($data, 'shipto_def', 'integer'))
@@ -106,7 +114,10 @@ class Address
             ->setCity(SHOP_getVar($data, 'city'))
             ->setState(SHOP_getVar($data, 'state'))
             ->setPostal(SHOP_getVar($data, 'zip'))
-            ->setCountry(SHOP_getVar($data, 'country', 'string', $_SHOP_CONF['country']));
+            ->setCountry(SHOP_getVar($data, 'country', 'string', $_SHOP_CONF['country']))
+            ->setBilltoDefault(SHOP_getVar($data, 'billto_def', 'integer'))
+            ->setShiptoDefault(SHOP_getVar($data, 'shipto_def', 'integer'));
+        return $this;
     }
 
 
@@ -122,19 +133,23 @@ class Address
         static $addrs = array();
 
         $addr_id = (int)$addr_id;
-        if (isset($addrs[$addr_id])) {
-            return new self($addrs[$addr_id]);
-        } else {
-            $res = DB_query("SELECT *
-                FROM {$_TABLES['shop.address']}
-                WHERE addr_id = '{$addr_id}'");
-            if ($res) {
-                $A = DB_fetchArray($res, true);
-                $addrs[$addr_id] = $A;
-                return new self($A);
+        if ($addr_id > 0) {
+            if (isset($addrs[$addr_id])) {
+                return new self($addrs[$addr_id]);
             } else {
-                return new self;
+                $res = DB_query("SELECT *
+                    FROM {$_TABLES['shop.address']}
+                    WHERE addr_id = '{$addr_id}'");
+                if ($res) {
+                    $A = DB_fetchArray($res, true);
+                    $addrs[$addr_id] = $A;
+                    return new self($A);
+                } else {
+                    return new self;
+                }
             }
+        } else {
+            return new self;
         }
     }
 
@@ -701,6 +716,35 @@ class Address
 
 
     /**
+     * Edit an address record.
+     *
+     * @return  string  HTML for editing form
+     */
+    public function Edit()
+    {
+        $T = new \Template(__DIR__ . '/../templates');
+        $T->set_file('form', 'editaddress.thtml');
+        $T->set_var(array(
+            'addr_id' => $this->addr_id,
+            'uid' => $this->uid,
+            'name' => $this->name,
+            'company' => $this->company,
+            'address1' => $this->address1,
+            'address2' => $this->address2,
+            'city' => $this->city,
+            'state' => $this->state,
+            'zip' => $this->zip,
+            'country_options' => Country::optionList($this->country),
+            'state_options' => State::optionList($this->country, $this->state),
+            'def_shipto_chk' => $this->isDefaultShipto() ? 'checked="checked"' : '',
+            'def_billto_chk' => $this->isDefaultBillto() ? 'checked="checked"' : '',
+        ) );
+        $T->parse('output', 'form');
+        return  $T->finish($T->get_var('output'));
+    }
+
+
+    /**
      * Save the address to the database.
      *
      * @return  integer     Record ID of address, zero on error
@@ -708,6 +752,11 @@ class Address
     public function Save()
     {
         global $_TABLES;
+
+        if ($this->uid < 2) {
+            // Got an invalid user ID, don't save.
+            return 0;
+        }
 
         if ($this->addr_id > 0) {
             $sql1 = "UPDATE {$_TABLES['shop.address']} SET ";
@@ -741,16 +790,19 @@ class Address
             // If this is the new default address, turn off the other default
             foreach (array('billto', 'shipto') as $type) {
                 if ($this->isDefault($type)) {
-                    DB_query(
-                        "UPDATE {$_TABLES['shop.address']}
-                        SET {$type}_def = 0
-                        WHERE addr_id <> '" . $this->addr_id . "' AND {$type}_def = 1"
-                    );
+                    $sql = "UPDATE {$_TABLES['shop.address']}
+                        SET {$type}_def = 0 WHERE 
+                        uid = {$this->uid}
+                        AND addr_id <> {$this->addr_id}
+                        AND {$type}_def = 1";
+                    DB_query($sql);
                 }
             }
+            Cache::clear('shop.user_' . $this->uid);
+            return $this->addr_id;
+        } else {
+            return 0;
         }
-        Cache::clear('shop.user_' . $this->uid);
-        return $this->addr_id;
     }
 
 
@@ -761,6 +813,7 @@ class Address
      */
     public function toArray()
     {
+        COM_errorLog("User ID: " . $this->uid);
         return array(
             'addr_id'   => $this->addr_id,
             'uid'       => $this->uid,
@@ -771,7 +824,9 @@ class Address
             'city'      => $this->city,
             'state'     => $this->state,
             'zip'       => $this->zip,
-            'country'       => $this->country,
+            'country'   => $this->country,
+            'billto_def' => $this->isDefaultBillto(),
+            'shipto_def' => $this->isDefaultShipto(),
         );
     }
 
@@ -890,8 +945,8 @@ class Address
             ->setCountry($Addr->getCountry())
             ->setPhone($Addr->getPhone());
         if ($all) {
-            $this->setDefaultBillto($Addr->isDefaultBillto())
-                ->setDefaultShipto($Addr->isDefaultShipto())
+            $this->setBilltoDefault($Addr->isDefaultBillto())
+                ->setShiptoDefault($Addr->isDefaultShipto())
                 ->setUid($Addr->getUid());
         }
         return $this;