+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## v1.3.0
 Release TBD
+  * Remove `paid` as an order status, check total paid instead.
+  * Update Square API to 3.20200325.0.
   * Add interface for customers to edit addresses outside of orders.
   * Add fast-checkout where the order allows it.
-  * Add invoicing via Paypal.
   * Enable sales tax charges on shipping and handling.
   * Add invoicing via Paypal invoice API.
   * Link to customer in order report now filters report.